refactor renaming of ParserUtils to parserUtils step 2
