+++ test/unit/MutantSpec.ts
@@ -3,7 +3,7 @@
 var expect = require('chai').expect;
 import Mutant, {MutantStatus} from '../../src/Mutant';
 import MathMutation from '../../src/mutations/MathMutation';
-import * as parserUtils from '../../src/utils/ParserUtils';
+import * as parserUtils from '../../src/utils/parserUtils';
 require('mocha-sinon');
 
 describe('Mutant', function() {
